Move binding types to top-level
This separates the implementations from abstract types in preparation for refactoring and multibinds
