+++ reflect/src/main/java/dagger/reflect/Binding.java
@@ -15,21 +15,7 @@
  */
 package dagger.reflect;
 
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.lang.reflect.Type;
-import java.util.Optional;
-import javax.inject.Inject;
 import javax.inject.Provider;
-
-import org.jetbrains.annotations.Nullable;
-
-import static dagger.reflect.DaggerReflect.notImplemented;
-import static dagger.reflect.Reflection.findQualifier;
-import static dagger.reflect.Reflection.tryInstantiate;
-import static dagger.reflect.Reflection.tryInvoke;
 
 interface Binding {
   abstract class UnlinkedBinding implements Binding {
@@ -55,212 +41,5 @@ interface Binding {
 
   abstract class LinkedBinding<T> implements Binding, Provider<T> {
     @Override public abstract String toString();
-  }
-
-  final class Instance<T> extends LinkedBinding<T> {
-    private final T value;
-
-    Instance(T value) {
-      this.value = value;
-    }
-
-    @Override public T get() {
-      return value;
-    }
-
-    @Override public String toString() {
-      return "@BindsInstance[" + value + ']';
-    }
-  }
-
-  final class UnlinkedBinds extends UnlinkedBinding {
-    private final Method method;
-
-    UnlinkedBinds(Method method) {
-      this.method = method;
-    }
-
-    @Override public LinkRequest request() {
-      Type[] parameterTypes = method.getGenericParameterTypes();
-      if (parameterTypes.length != 1) {
-        throw new IllegalArgumentException(
-            "@Binds methods must have a single parameter: " + method);
-      }
-      Annotation[][] parameterAnnotations = method.getParameterAnnotations();
-      Key dependency = Key.of(findQualifier(parameterAnnotations[0]), parameterTypes[0]);
-      return new LinkRequest(new Key[] { dependency });
-    }
-
-    @Override public LinkedBinding<?> link(LinkedBinding<?>[] dependencies) {
-      return dependencies[0];
-    }
-  }
-
-  final class UnlinkedOptionalBinding extends UnlinkedBinding {
-    private final Method method;
-
-    UnlinkedOptionalBinding(Method method) {
-      this.method = method;
-    }
-
-    @Override
-    public LinkRequest request() {
-      Type[] parameterTypes = method.getGenericParameterTypes();
-      if (parameterTypes.length != 0) {
-        throw new IllegalArgumentException(
-            "@BindsOptionalOf methods must not have parameters: " + method);
-      }
-
-      Annotation[] methodAnnotations = method.getDeclaredAnnotations();
-      Annotation qualifier = findQualifier(methodAnnotations);
-      Key dependency = Key.of(qualifier, method.getReturnType());
-      return new LinkRequest(new Key[] { dependency }, new boolean[] { true });
-    }
-
-    @Override
-    public LinkedBinding<?> link(LinkedBinding<?>[] dependencies) {
-      return new LinkedOptionalBinding<>(dependencies[0]);
-    }
-  }
-
-  final class LinkedOptionalBinding<T> extends LinkedBinding<Optional<T>> {
-    private final @Nullable LinkedBinding<T> dependency;
-
-    LinkedOptionalBinding(@Nullable LinkedBinding<T> dependency) {
-      this.dependency = dependency;
-    }
-
-    @Override
-    public Optional<T> get() {
-      if (dependency == null) {
-        return Optional.empty();
-      }
-      T value = dependency.get();
-      if (value == null) {
-        throw new NullPointerException(
-            dependency + " returned null which is not allowed for optional bindings");
-      }
-      return Optional.of(value);
-    }
-
-    @Override public String toString() {
-      return "@BindsOptionalOf[" + dependency + ']';
-    }
-  }
-
-  final class UnlinkedProvides extends UnlinkedBinding {
-    private final @Nullable Object instance;
-    private final Method method;
-
-    UnlinkedProvides(@Nullable Object instance, Method method) {
-      this.instance = instance;
-      this.method = method;
-    }
-
-    @Override public LinkRequest request() {
-      Type[] parameterTypes = method.getGenericParameterTypes();
-      if (parameterTypes.length == 0) {
-        return LinkRequest.EMPTY;
-      }
-      Annotation[][] parameterAnnotations = method.getParameterAnnotations();
-      Key[] dependencies = new Key[parameterTypes.length];
-      for (int i = 0; i < parameterTypes.length; i++) {
-        dependencies[i] = Key.of(findQualifier(parameterAnnotations[i]), parameterTypes[i]);
-      }
-      return new LinkRequest(dependencies);
-    }
-
-    @Override public LinkedBinding<?> link(LinkedBinding<?>[] dependencies) {
-      return new LinkedProvides<>(instance, method, dependencies);
-    }
-  }
-
-  final class LinkedProvides<T> extends LinkedBinding<T> {
-    private final @Nullable Object instance;
-    private final Method method;
-    private final LinkedBinding<?>[] dependencies;
-
-    LinkedProvides(@Nullable Object instance, Method method, LinkedBinding<?>[] dependencies) {
-      this.instance = instance;
-      this.method = method;
-      this.dependencies = dependencies;
-    }
-
-    @Override public @Nullable T get() {
-      Object[] arguments = new Object[dependencies.length];
-      for (int i = 0; i < arguments.length; i++) {
-        arguments[i] = dependencies[i].get();
-      }
-      // The binding is associated with the return type of method as key.
-      @SuppressWarnings("unchecked")
-      T value = (T) tryInvoke(instance, method, arguments);
-      return value;
-    }
-
-    @Override public String toString() {
-      return "@Provides[" + method.getDeclaringClass().getName() + '.' + method.getName() + "(…)]";
-    }
-  }
-
-  final class UnlinkedJustInTime<T> extends UnlinkedBinding {
-    private final Class<T> cls;
-    private final Constructor<T> constructor;
-
-    UnlinkedJustInTime(Class<T> cls, Constructor<T> constructor) {
-      this.cls = cls;
-      this.constructor = constructor;
-    }
-
-    @Override public LinkRequest request() {
-      // TODO field and method bindings? reuse some/all of reflective members injector somehow?
-      Class<?> target = cls;
-      while (target != Object.class) {
-        for (Field field : target.getDeclaredFields()) {
-          if (field.getAnnotation(Inject.class) != null) {
-            throw notImplemented("@Inject on fields in just-in-time bindings");
-          }
-        }
-        for (Method method : target.getDeclaredMethods()) {
-          if (method.getAnnotation(Inject.class) != null) {
-            throw notImplemented("@Inject on methods in just-in-time bindings");
-          }
-        }
-        target = target.getSuperclass();
-      }
-
-      Type[] parameterTypes = constructor.getGenericParameterTypes();
-      Annotation[][] parameterAnnotations = constructor.getParameterAnnotations();
-      Key[] dependencies = new Key[parameterTypes.length];
-      for (int i = 0; i < parameterTypes.length; i++) {
-        dependencies[i] = Key.of(findQualifier(parameterAnnotations[i]), parameterTypes[i]);
-      }
-      return new LinkRequest(dependencies);
-    }
-
-    @Override public LinkedBinding<T> link(LinkedBinding<?>[] dependencies) {
-      return new LinkedJustInTime<>(constructor, dependencies);
-    }
-  }
-
-  final class LinkedJustInTime<T> extends LinkedBinding<T> {
-    private final Constructor<T> constructor;
-    private final LinkedBinding<?>[] dependencies;
-
-    LinkedJustInTime(Constructor<T> constructor, LinkedBinding<?>[] dependencies) {
-      this.constructor = constructor;
-      this.dependencies = dependencies;
-    }
-
-    @Override public T get() {
-      Object[] arguments = new Object[dependencies.length];
-      for (int i = 0; i < dependencies.length; i++) {
-        arguments[i] = dependencies[i].get();
-      }
-      return tryInstantiate(constructor, arguments);
-    }
-
-    @Override public String toString() {
-      return "@Inject[" + constructor.getDeclaringClass().getName() + ".<init>(…)]";
-    }
   }
 }
